% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.health_operations.R
\name{describe_event_aggregates}
\alias{describe_event_aggregates}
\title{Returns the number of events of each event type (issue, scheduled change, and account notification)}
\usage{
describe_event_aggregates(aggregateField, filter = NULL, maxResults = NULL,
  nextToken = NULL)
}
\arguments{
\item{aggregateField}{The only currently supported value is \code{eventTypeCategory}.}

\item{filter}{Values to narrow the results returned.}

\item{maxResults}{The maximum number of items to return in one batch, between 10 and 100, inclusive.}

\item{nextToken}{If the results of a search are large, only a portion of the results are returned, and a \code{nextToken} pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.}
}
\description{
Returns the number of events of each event type (issue, scheduled change, and account notification). If no filter is specified, the counts of all events in each category are returned.
}
\examples{

}
