% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.health_operations.R
\name{describe_event_details}
\alias{describe_event_details}
\title{Returns detailed information about one or more specified events}
\usage{
describe_event_details(eventArns, locale = NULL)
}
\arguments{
\item{eventArns}{[required] A list of event ARNs (unique identifiers). For example: \code{"arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"}}

\item{locale}{The locale (language) to return information in. English (en) is the default and the only supported value at this time.}
}
\description{
Returns detailed information about one or more specified events. Information includes standard event data (region, service, etc., as returned by DescribeEvents), a detailed event description, and possible additional metadata that depends upon the nature of the event. Affected entities are not included; to retrieve those, use the DescribeAffectedEntities operation.
}
\details{
If a specified event cannot be retrieved, an error message is returned for that event.
}
\section{Accepted Parameters}{
\preformatted{describe_event_details(
  eventArns = list(
    "string"
  ),
  locale = "string"
)
}
}

