% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.health_package.R
\docType{package}
\name{paws.health-package}
\alias{paws.health}
\alias{paws.health-package}
\title{paws.health: AWS Health APIs and Notifications}
\description{
- DescribeEvents: Summary information about
events.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
