% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.health_package.R
\docType{package}
\name{paws.health-package}
\alias{paws.health}
\alias{paws.health-package}
\title{paws.health: AWS Health APIs and Notifications}
\description{
The AWS Health API provides programmatic access to the AWS
Health information that is presented in the [AWS Personal Health
Dashboard](https://phd. aws. amazon. com/phd/home#/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
